[refactor] #27 post, comment 컨트롤러 분리 및 @ApiResponse 내용 추가
